requirements.txt
# Baby-Noise Generator v1.1 - Requirements
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
soundfile>=0.10.0
sounddevice>=0.4.5
matplotlib>=3.5.0
pyyaml>=5.1.0
pyloudnorm>=0.1.0
pytest>=7.0.0

# Optional GPU acceleration
# cupy>=12.0.0
# cupyx-scipy>=12.0.0